+++ packages/elements/emoji/src/typeahead.ts
@@ -1,0 +1,3 @@
+export {
+  default as EmojiTypeAhead,
+} from './components/typeahead/EmojiTypeAhead';