+++ packages/media/media-viewer/src/newgen/utils/getIdentifierCollection.ts
@@ -1,0 +1,9 @@
+import { isFileIdentifier, Identifier } from '@atlaskit/media-core';
+
+export const getIdentifierCollection = (
+  identifier: Identifier,
+  defaultCollectionName: string,
+): string | undefined =>
+  isFileIdentifier(identifier)
+    ? identifier.collectionName || defaultCollectionName
+    : undefined;