+++ packages/search/global-search/example-helpers/BasicNavigation.tsx
@@ -19,7 +19,7 @@ export interface State {
 }
 
 export default class BasicNavigation extends React.Component<Props, State> {
-  constructor(props) {
+  constructor(props: Props) {
     super(props);
     this.state = {
       openDrawer: 'search',