+++ packages/core/global-navigation/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @atlaskit/global-navigation
 
+## 6.2.5
+- Updated dependencies [e36f791fd6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/e36f791fd6):
+  - @atlaskit/atlassian-switcher@0.4.0
+
+## 6.2.4
+- [patch] [db2a7ffde6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/db2a7ffde6):
+
+  - Fixing recent containers bug
+
+## 6.2.3
+- [patch] [9d6f8d516a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/9d6f8d516a):
+
+  - Adding expand link support to Atlassian Switcher
+
 ## 6.2.2
 - Updated dependencies [987ab01f30](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/987ab01f30):
 - Updated dependencies [c95557e3ff](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c95557e3ff):